Fixing problem
When I was done with the previous problem I went on to a new task and it
was to create a new menu after the user has created an account and wants
to login. I saw a new problem and its after the new account has been
added my menu doesnt come up again. I have to find a way to move all the
console.writelines between the booleans so I can get it right.