Whole task is completed and Im very happy.
After I completed my case "5" when it comes to accept or deny a request.
I felt that something is missing. When a user login I would like to have
a integer as a notification beside the menu choice number 5 in the end.
Example: (1). Now when Im done with everything I would like to try
implement StreamReader to my program and after that I will try to implement the notification.