Everything is completed.
After many days I finally did it. A very difficult task but i've learned
so so so much. I understand the logic behind the whole program but I
still find it very hard to put it down in codes which I guess will
become easier and easier during this education. In the end I even
implemented a notification in the menu choice "5" by easly implement a
integer "waiting" which loops through all the trades and with a
if-statement in the begining of the second menu-loop. The last thing I
like to do is implement the streamreader and writer part which I also
find very difficult. I need to find a little bit more energy and
motivation because the developing of my trading system has taking a lot
of my energy.