Mission completed
I've reached the goal except the StreamReader and StreamWrite part which
I'm a little sad for because it was something that I really wanted my
program to have. But I'm guessing that I can learn more about it and
implement it tomorrow when we have our Q&A-moment. I have one readme
file with information about how the program works and I have one readme2
were you can read all about my reasons for the choices of different methods/functions that I chose to use in my system.